Move srg_utils to a module so the openpyxl is only included as needed
