+++ utils/srg_diff.py
@@ -7,9 +7,12 @@ import os.path
 import jinja2
 
 from openpyxl import load_workbook
+from openpyxl.worksheet.worksheet import Worksheet
+from srg_utils.worksheet_utils import Row, get_stigid_set, get_cce_dict_to_row_dict
+from srg_utils import get_rule_dir_json, get_cce_dict, get_full_name
 
 import ssg.jinja
-from utils.srg_utils import *
+
 
 SSG_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 BUILD_ROOT = os.path.join(SSG_ROOT, "build")
